Move some functions from checkmember to typeops
